requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Azure OpenAI
openai==1.12.0

# PDF Processing
pypdf==4.0.1

# Progress Tracking
tqdm==4.66.1

# Azure AI Search
azure-search-documents==11.4.0
azure-identity==1.15.0

# Utilities
tiktoken==0.5.2

